+++ CommunalCalculatorTest/Models/Services/Imp/ThermalEnergyByMeter.cs
@@ -2,8 +2,8 @@
 {
     internal class ThermalEnergyByMeter : CommunalServiceByMeter
     {
-        public ThermalEnergyByMeter(int scopeOfService) : base(scopeOfService) { }
-        public ThermalEnergyByMeter(int readingBefore, int readingNow) : base(readingBefore, readingNow) { }
+        public ThermalEnergyByMeter(decimal scopeOfService) : base(scopeOfService) { }
+        public ThermalEnergyByMeter(decimal readingBefore, decimal readingNow) : base(readingBefore, readingNow) { }
         internal override decimal GetSalary()
         {
             return this._scopeOfServices * this.Rate.Normative * this.Rate.Rate;
